[MLIR][AFFINE] Handling possible side-effects during fusion
Check the fusion recommended forOps for possible side-effects. This commit assummes is, if there is a function call present in the forOp, it may have possible side-effects and thus, fusion is not possible. Since, the loop fusion strategies are based on common memref access, pressence of possible side-effects in either of the loops voilates the precondition for loop fusion.